Done with the final endpoint

+++ src/controllers/post.controller.js
@@ -1,9 +1,13 @@
 const Post = require('../models/Post');
 const PostImage = require('../models/PostImage');
-const { newPostValidation } = require('../utils/validation');
+const {
+  newPostValidation,
+  editPostValidation,
+} = require('../utils/validation');
 
 const validationObject = {
   newPost: newPostValidation,
+  editPost: editPostValidation,
 };
 
 const handleValidation = (body, type) => {
@@ -14,8 +18,8 @@ const handleValidation = (body, type) => {
 };
 const composePostResponse = async (post, imageIds) => {
   const images = await PostImage.find({ _id: imageIds });
-  const { content, _id } = post;
-  const responseData = { content, _id, images };
+  const { content, _id, user } = post;
+  const responseData = { content, _id, images, user };
   return responseData;
 };
 const createPostController = async (req, res) => {
@@ -24,6 +28,7 @@ const createPostController = async (req, res) => {
 
     await handleValidation(req.body, 'newPost');
     const newPost = new Post(req.body);
+    newPost.user = req.user._id;
     await newPost.save();
 
     return res.status(200).json({
@@ -57,8 +62,44 @@ const getASinglePostController = async (req, res) => {
 
 const deletePostController = async (req, res) => {
   try {
+    if (String(req.user._id) !== String(post.user)) {
+      return res
+        .status(401)
+        .json({ error: `You don't have permission to delete this post` });
+    }
     await Post.findOneAndDelete({ _id: req.params.id });
     return res.status(200).json({ success: true, message: 'Deleted success' });
+  } catch (error) {
+    return res.status(400).json({ error: error.message });
+  }
+};
+
+const editPostController = async (req, res) => {
+  try {
+    await handleValidation(req.body, 'editPost');
+    const post = await Post.findById(req.params.id);
+    if (!post) {
+      return res.status(404).json({ error: 'Post with that ID not found' });
+    }
+    if (String(req.user._id) !== String(post.user)) {
+      return res
+        .status(401)
+        .json({ error: `You don't have permission to edit this post` });
+    }
+    const updatedPost = await Post.findByIdAndUpdate(
+      { _id: req.params.id },
+      req.body,
+      {
+        new: true,
+        runValidators: true,
+      }
+    );
+    const { imageIds } = updatedPost;
+    return res.status(200).json({
+      success: true,
+      message: 'Edited success',
+      data: await composePostResponse(updatedPost, imageIds),
+    });
   } catch (error) {
     return res.status(400).json({ error: error.message });
   }
@@ -68,4 +109,5 @@ module.exports = {
   createPostController,
   getASinglePostController,
   deletePostController,
+  editPostController,
 };
